Integrating Joachim's changes requests
