+++ src/Client/Client.php
@@ -134,6 +134,12 @@ final class Client implements ClientInterface
      */
     public function updateMember(AudienceInterface $audience, CustomerInterface $customer): void
     {
+        Assert::notNull($customer->getEmail());
+
+        if (null === $customer->getFirstName() || null === $customer->getLastName()) {
+            $this->subscribeEmail($audience, $customer->getEmail());
+        }
+
         $data = [
             'email_address' => $customer->getEmail(),
             'status' => 'subscribed',
@@ -143,8 +149,6 @@ final class Client implements ClientInterface
                 'LNAME' => $customer->getLastName(),
             ],
         ];
-
-        Assert::notNull($customer->getEmail());
 
         $this->makeRequest('put',
             sprintf(
