+++ src/Message/Command/PushCustomer.php
@@ -9,22 +9,13 @@ final class PushCustomer implements CommandInterface
     /** @var int */
     private $customerId;
 
-    /** @var bool */
-    private $pushEmailOnly;
-
-    public function __construct(int $customerId, bool $pushEmailOnly = false)
+    public function __construct(int $customerId)
     {
         $this->customerId = $customerId;
-        $this->pushEmailOnly = $pushEmailOnly;
     }
 
     public function getCustomerId(): int
     {
         return $this->customerId;
-    }
-
-    public function isPushEmailOnly(): bool
-    {
-        return $this->pushEmailOnly;
     }
 }
