+++ src/Message/Handler/PushCustomerHandler.php
@@ -52,7 +52,7 @@ final class PushCustomerHandler implements MessageHandlerInterface
         $audienceId = $audience->getId();
         Assert::notNull($audienceId);
 
-        $subscribeCustomerMessage = new SubscribeCustomerToAudience($audienceId, $customer->getId(), $message->isPushEmailOnly());
+        $subscribeCustomerMessage = new SubscribeCustomerToAudience($audienceId, $customer->getId());
         $this->messageBus->dispatch($subscribeCustomerMessage);
     }
 }
