--- src/Resources/config/services/entity_listener.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<container xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-           xmlns="http://symfony.com/schema/dic/services"
-           xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
-
-    <services>
-        <service id="setono_sylius_mailchimp.event_listener.update_store"
-                 class="Setono\SyliusMailchimpPlugin\EventListener\UpdateStoreSubscriber">
-            <argument type="service" id="setono_sylius_mailchimp.client"/>
-
-            <tag name="kernel.event_subscriber"/>
-        </service>
-
-        <service id="Setono\SyliusMailchimpPlugin\EntityListener\Customer\PushCustomerToMailchimp">
-            <argument type="service" id="setono_sylius_mailchimp.command_bus"/>
-
-            <tag name="doctrine.orm.entity_listener" event="postPersist" entity="%sylius.model.customer.class%" />
-            <tag name="doctrine.orm.entity_listener" event="postUpdate" entity="%sylius.model.customer.class%" />
-        </service>
-    </services>
-
-</container>
