+++ src/Resources/config/services/event_listener.xml
@@ -18,6 +18,13 @@
 
             <tag name="kernel.event_subscriber"/>
         </service>
+
+        <service id="Setono\SyliusMailchimpPlugin\EventListener\Doctrine\Customer\PushCustomerToMailchimp">
+            <argument type="service" id="setono_sylius_mailchimp.command_bus"/>
+
+            <tag name="doctrine.orm.entity_listener" event="postPersist" entity="%sylius.model.customer.class%" />
+            <tag name="doctrine.orm.entity_listener" event="postUpdate" entity="%sylius.model.customer.class%" />
+        </service>
     </services>
 
 </container>
